backend/config.py
import os
from dotenv import load_dotenv


basedir = os.path.abspath(os.path.dirname(__file__))
load_dotenv(os.path.join(basedir, '.env'))


class Config(object):
    SECRET_KEY = os.environ.get('SECRET_KEY') or 'super-pass'
    SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_URL') or \
        'sqlite:///' + os.path.join(basedir, 'app.db')  # ? sqlite example
    SQLALCHEMY_TRACK_MODIFICATIONS = False
    # JWT_ACCESS_LIFESPAN = {'minutes': 50}  # ? GUARD token lifespan


class DevelopmentConfig(Config):
    DEBUG = True
    SEND_FILE_MAX_AGE_DEFAULT = 0  # ? disable static file cache JS CSS
    JSON_SORT_KEYS = False
