backend/requirements.txt
alembic==1.4.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
flask-marshmallow==0.11.0
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.1
MarkupSafe==1.1.1
marshmallow==3.4.0
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
python-dateutil==2.8.1
python-dotenv==0.11.0
python-editor==1.0.4
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.13
urllib3==1.25.8
Werkzeug==1.0.0
